requirements.txt
modal==1.0.4
python-dotenv==1.1.0
openai==1.86.0